moved choose_pool to separate file
